+++ Code/Controllers/ATLConversationListViewController.m
@@ -24,6 +24,7 @@
 
 static NSString *const ATLConversationCellReuseIdentifier = @"ATLConversationCellReuseIdentifier";
 static NSString *const ATLImageMIMETypePlaceholderText = @"Attachment: Image";
+static NSString *const ATLVideoMIMETypePlaceholderText = @"Attachment: Video";
 static NSString *const ATLLocationMIMETypePlaceholderText = @"Attachment: Location";
 static NSString *const ATLGIFMIMETypePlaceholderText = @"Attachment: GIF";
 
@@ -84,6 +85,7 @@ NSString *const ATLConversationListViewControllerDeletionModeGlobal = @"Global";
     _displaysAvatarItem = NO;
     _allowsEditing = YES;
     _rowHeight = 76.0f;
+    _shouldDisplaySearchController = YES;
 }
 
 - (id)init
@@ -113,6 +115,7 @@ NSString *const ATLConversationListViewControllerDeletionModeGlobal = @"Global";
     self.tableView.isAccessibilityElement = YES;
     self.tableView.tableFooterView = [[UIView alloc] initWithFrame:CGRectZero];
     
+<<<<<<< HEAD
     /*self.searchBar = [[UISearchBar alloc] initWithFrame:CGRectZero];
     [self.searchBar sizeToFit];
     self.searchBar.translucent = NO;
@@ -120,13 +123,30 @@ NSString *const ATLConversationListViewControllerDeletionModeGlobal = @"Global";
     self.searchBar.delegate = self;
     self.tableView.tableHeaderView = self.searchBar;
     
+=======
+    if (self.shouldDisplaySearchController) {
+        self.searchBar = [[UISearchBar alloc] initWithFrame:CGRectZero];
+        [self.searchBar sizeToFit];
+        self.searchBar.translucent = NO;
+        self.searchBar.accessibilityLabel = @"Search Bar";
+        self.searchBar.delegate = self;
+        self.tableView.tableHeaderView = self.searchBar;
+        
+>>>>>>> layerhq/master
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wdeprecated-declarations"
-    self.searchController = [[UISearchDisplayController alloc] initWithSearchBar:self.searchBar contentsController:self];
+        self.searchController = [[UISearchDisplayController alloc] initWithSearchBar:self.searchBar contentsController:self];
 #pragma GCC diagnostic pop
+<<<<<<< HEAD
     self.searchController.delegate = self;
     self.searchController.searchResultsDelegate = self;
     self.searchController.searchResultsDataSource = self;*/
+=======
+        self.searchController.delegate = self;
+        self.searchController.searchResultsDelegate = self;
+        self.searchController.searchResultsDataSource = self;
+    }
+>>>>>>> layerhq/master
 }
 
 - (void)viewWillAppear:(BOOL)animated
@@ -514,6 +534,8 @@ NSString *const ATLConversationListViewControllerDeletionModeGlobal = @"Global";
             lastMessageText = ATLLocalizedString(@"atl.conversationlist.lastMessage.text.gif.key", ATLGIFMIMETypePlaceholderText, nil);
         } else if ([messagePart.MIMEType isEqualToString:ATLMIMETypeLocation]) {
             lastMessageText = ATLLocalizedString(@"atl.conversationlist.lastMessage.text.location.key", ATLLocationMIMETypePlaceholderText, nil);
+        } else if ([messagePart.MIMEType isEqualToString:ATLMIMETypeVideoMP4]) {
+            lastMessageText = ATLLocalizedString(@"atl.conversationlist.lastMessage.text.video.key", ATLVideoMIMETypePlaceholderText, nil);
         } else {
             lastMessageText = ATLLocalizedString(@"atl.conversationlist.lastMessage.text.default.key", ATLImageMIMETypePlaceholderText, nil);
         }