+++ Code/Utilities/ATLMessagingUtilities.h
@@ -19,10 +19,12 @@
 //
 
 #import <Foundation/Foundation.h>
-#import <LayerKit/LayerKit.h>
+@import LayerKit;
 #import <MapKit/MapKit.h>
 #import <ImageIO/ImageIO.h>
 #import "ATLMediaAttachment.h"
+#import "UIResponder+ATLFirstResponder.h"
+#import "ATLMessageComposeTextView.h"
 
 extern NSString *const ATLMIMETypeTextPlain;          // text/plain
 extern NSString *const ATLMIMETypeImagePNG;           // image/png
@@ -31,9 +33,10 @@ extern NSString *const ATLMIMETypeImageJPEGPreview;   // image/jpeg+preview
 extern NSString *const ATLMIMETypeImageGIF;           // image/gif
 extern NSString *const ATLMIMETypeImageGIFPreview;    // image/gif+preview
 extern NSString *const ATLMIMETypeImageSize;          // application/json+imageSize
+extern NSString *const ATLMIMETypeVideoQuickTime;     // video/quicktime
 extern NSString *const ATLMIMETypeLocation;           // location/coordinate
 extern NSString *const ATLMIMETypeDate;               // text/date
-
+extern NSString *const ATLMIMETypeVideoMP4;           // video/mp4
 extern NSUInteger const ATLDefaultThumbnailSize;      // 512px
 extern NSUInteger const ATLDefaultGIFThumbnailSize;   // 64px
 
@@ -78,6 +81,14 @@ CGSize ATLTextPlainSize(NSString *string, UIFont *font);
 
 CGRect ATLImageRectConstrainedToSize(CGSize imageSize, CGSize maxSize);
 
+CGFloat ATLDegreeToRadians(CGFloat degrees);
+
+//------------------------
+// @name Message Utilities
+//------------------------
+
+LYRMessage *ATLMessageForParts(LYRClient *layerClient, NSArray *messageParts, NSString *pushText, NSString *pushSound);
+
 //-----------------------------
 // @name Message Part Utilities
 //-----------------------------
@@ -97,3 +108,9 @@ void ATLLastPhotoTaken(void(^completionHandler)(UIImage *image, NSError *error))
 UIImage *ATLPinPhotoForSnapshot(MKMapSnapshot *snapshot, CLLocationCoordinate2D location);
 
 NSArray *ATLTextCheckingResultsForText(NSString *text, NSTextCheckingType linkTypes);
+
+//---------------------------------
+// @name Resources Bundle Utilities
+//---------------------------------
+
+NSBundle *ATLResourcesBundle(void);