+++ Code/Utilities/UIResponder+ATLFirstResponder.m
@@ -1,8 +1,8 @@
 //
-//  ATLLayerKitMock.h
+//  UIResponder+ATLFirstResponder.m
 //  Atlas
 //
-//  Created by Kevin Coleman on 12/8/14.
+//  Created by Kabir Mahal on 5/15/15.
 //  Copyright (c) 2015 Layer. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,14 +18,20 @@
 //  limitations under the License.
 //
 
-#import <Atlas/Atlas.h>
+#import "UIResponder+ATLFirstResponder.h"
 
-#import "LYRClientMock.h"
-#import "LYRConversationMock.h"
-#import "LYRMessageMock.h"
-#import "LYRMessagePartMock.h"
+static __weak id currentFirstResponder;
 
-#import "LYRQueryControllerMock.h"
+@implementation UIResponder (ATLFirstResponder)
 
-#import "LYRMockContentStore.h"
-#import "ATLUserMock.h"
++ (id)atl_currentFirstResponder {
+    currentFirstResponder = nil;
+    [[UIApplication sharedApplication] sendAction:@selector(atl_findFirstResponder:) to:nil from:nil forEvent:nil];
+    return currentFirstResponder;
+}
+
+- (void)atl_findFirstResponder:(id)sender {
+    currentFirstResponder = self;
+}
+
+@end