+++ Code/Views/ATLMessageBubbleView.h
@@ -24,6 +24,7 @@
 
 extern CGFloat const ATLMessageBubbleLabelVerticalPadding;
 extern CGFloat const ATLMessageBubbleLabelHorizontalPadding;
+extern CGFloat const ATLMessageBubbleLabelWidthMargin;
 extern CGFloat const ATLMessageBubbleMapWidth;
 extern CGFloat const ATLMessageBubbleMapHeight;
 extern CGFloat const ATLMessageBubbleDefaultHeight;
@@ -55,6 +56,11 @@ extern NSString *const ATLUserDidTapPhoneNumberNotification;
  @abstract Tells the bubble view to display a given image.
  */
 - (void)updateWithImage:(UIImage *)image width:(CGFloat)width;
+
+/**
+ @abstract Tells the bubble view to display the thumbnail for a video.
+ */
+- (void)updateWithVideoThumbnail:(UIImage *)image width:(CGFloat)width;
 
 /**
  @abstract Tells the bubble view to display a map image for a given location.