+++ Code/Views/ATLMessageInputToolbar.h
@@ -97,9 +97,14 @@ extern NSString *const ATLMessageInputToolbarAccessibilityLabel;
  
 /**
  @abstract The right accessory button for the view.
- @discussion By default, the button displays the text "SEND".
  */
 @property (nonatomic) UIButton *rightAccessoryButton;
+
+/**
+ @abstract The right accessory button title.
+ @discussion By default, the title is "Send".
+ */
+@property (nonatomic) NSString *rightAccessoryButtonTitle;
 
 /**
  @abstract The font color for the right accessory button in active state.
@@ -141,6 +146,12 @@ extern NSString *const ATLMessageInputToolbarAccessibilityLabel;
 @property (nonatomic) ATLMessageComposeTextView *textInputView;
 
 /**
+  @abstract The margin on top and bottom of the textInputView.
+  @default 7.0f.
+  */
+@property (nonatomic) CGFloat verticalMargin;
+
+/**
  @abstract The delegate object for the view.
  */
 @property (nonatomic, weak) id<ATLMessageInputToolbarDelegate> inputToolBarDelegate;
@@ -158,5 +169,16 @@ extern NSString *const ATLMessageInputToolbarAccessibilityLabel;
  @discussion Any existing media attachments will be removed when the right accessory button is tapped.
  */
 @property (nonatomic, readonly) NSArray *mediaAttachments;
+
+//-------------------
+// Layout Accessories
+//-------------------
+
+/**
+ @abstract The view controller whose input accessory view is the `ATLMessageInputToolbar`.
+ @discussion This property is set internally in the `ATLBaseConversationViewController` to change the view's frame
+ to support UISplitViewController usage.  This property should only be set when subclassing `ATLMessageInputToolbar`.
+ */
+@property (nonatomic, weak) UIViewController *containerViewController;
 
 @end